backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiofiles>=23.2.1
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
torch>=2.0.0
torchvision>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
